Added comments to provide more detail.

+++ MovingRobot/src/main/java/GameConstraints/Line.java
@@ -2,6 +2,14 @@ package GameConstraints;
 
 import java.awt.Point;
 
+/**
+ * Line class used for tracking collisions. 
+ * Stores lines of the form:
+ * <p>ax + by + c = 0</p>
+ * Between Points (x1, y1) and (x2, y2).
+ * 
+ * @author Kevin Caldwell
+ */
 public class Line {
     public double a = 0;
     public double b = 0;
@@ -11,13 +19,26 @@ public class Line {
     public double x2 = 0;
     public double y2 = 0;
 
-    public Line(int a, int b, int c) {
+    /**
+     * Constructor which takes a, b, c as inputs
+     * @param a
+     * @param b
+     * @param c
+     */
+    public Line(double a, double b, double c) {
         this.a = a;
         this.b = b;
         this.c = c;
     }
 
-    public Line(int x1, int y1, int x2, int y2) {
+    /**
+     * Contructor with 2 Points, (x1, y1) and (x2, y2) 
+     * @param x1
+     * @param y1
+     * @param x2
+     * @param y2
+     */
+    public Line(double x1, double y1, double x2, double y2) {
         this.x1 = x1;
         this.x2 = x2;
         this.y1 = y1;
@@ -28,6 +49,11 @@ public class Line {
         c = x1 * y2 - x2 * y1;
     }
 
+    /**
+     * Contructor with 2 Points, p1 and p2
+     * @param p1
+     * @param p2
+     */
     public Line(Point p1, Point p2) {
         x1 = p1.getX();
         y1 = p1.getY();
